requirements.txt
Flask
psycopg2-binary
python-dotenv
requests
requests
flask_login
flask_sqlalchemy==3.1.1
